Add toggles to benchmark details page
